refactor: 로그인/회원가입/메인화면 리팩토링 및 부가 기능 추가
- 디자인 디테일 조정
- viewModel/UseCase/Repository 일부 리팩토링
- 메인 화면 사진 기능 추가
