+++ app/src/main/java/com/kau/ttokttok/data/remote/api/AuthApiService.kt
@@ -3,9 +3,12 @@ package com.kau.ttokttok.data.remote.api
 import com.kau.ttokttok._core.network.model.ApiResponse
 import com.kau.ttokttok.data.remote.dto.auth.req.LoginReq
 import com.kau.ttokttok.data.remote.dto.auth.req.RegisterReq
+import com.kau.ttokttok.data.remote.dto.auth.req.RequestTempPasswordReq
+import com.kau.ttokttok.data.remote.dto.auth.req.VerifyEmailReq
 import com.kau.ttokttok.data.remote.dto.auth.res.LoginRes
 import com.kau.ttokttok.data.remote.dto.auth.res.RegisterRes
 import retrofit2.http.Body
+import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.POST
 
@@ -17,4 +20,12 @@ interface AuthApiService {
     @Headers("No-Auth: true")
     @POST("api/users/join")
     suspend fun register(@Body req: RegisterReq): ApiResponse<RegisterRes>
+
+    @Headers("No-Auth: true")
+    @POST("api/users/issue-temporary-password")
+    suspend fun requestTemporaryPassword(@Body req: RequestTempPasswordReq): ApiResponse<String>
+
+    @Headers("api/users/verify-email")
+    @GET("api/users/verify-email")
+    suspend fun verifyEmail(@Body req: VerifyEmailReq): ApiResponse<String>
 }
