+++ app/src/main/java/com/kau/ttokttok/data/remote/dto/auth/req/RequestTempPasswordReq.kt
@@ -1,0 +1,5 @@
+package com.kau.ttokttok.data.remote.dto.auth.req
+
+data class RequestTempPasswordReq(
+    val email: String
+)
