+++ app/src/main/java/com/kau/ttokttok/data/remote/dto/auth/req/VerifyEmailReq.kt
@@ -1,0 +1,5 @@
+package com.kau.ttokttok.data.remote.dto.auth.req
+
+data class VerifyEmailReq(
+    val email: String
+)
