+++ app/src/main/java/com/kau/ttokttok/ui/component/common/card/GlassCard.kt
@@ -11,68 +11,24 @@ import androidx.compose.material3.CardDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun GlassCard(
-    modifier: Modifier = Modifier,
-    shape: Shape = RoundedCornerShape(16.dp),
-    elevation: Dp = 3.dp,
-    containerAlpha: Float = 0.08f,   // 카드 내부 유리감
-    borderAlpha: Float = 0.12f,      // 외곽선 투명도
-    shadowAlpha: Float = 0.15f,      // 그림자 투명도
-    shadowColor: Color = Color(0xFF1A237E),
-    padding: PaddingValues = PaddingValues(16.dp),
-    content: @Composable ColumnScope.() -> Unit
-) {
-    Card(
-        modifier = modifier
-            .clip(shape) // 사각 경계 이슈 방지
-            .shadow(
-                elevation = elevation,
-                shape = shape,
-                ambientColor = Color.Black.copy(alpha = shadowAlpha),
-                spotColor = Color.Black.copy(alpha = shadowAlpha)
-            )
-            .border(1.dp, Color.White.copy(alpha = borderAlpha), shape),
-        shape = shape,
-        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = Color.White.copy(alpha = containerAlpha)
-        )
-    ) {
-        Column(Modifier.padding(padding)) {
-            content()
-        }
-    }
-}
-
-@Composable
 fun GlassCardClickable(
-    onClick: () -> Unit,
     modifier: Modifier = Modifier,
+    onClick: () -> Unit = {},
     shape: Shape = RoundedCornerShape(16.dp),
-    elevation: Dp = 8.dp,
     containerAlpha: Float = 0.08f,
     borderAlpha: Float = 0.15f,
-    shadowAlpha: Float = 0.25f,
     padding: PaddingValues = PaddingValues(16.dp),
-    content: @Composable ColumnScope.() -> Unit
+    content: @Composable ColumnScope.() -> Unit = {}
 ) {
     Card(
         onClick = onClick,
         modifier = modifier
             .clip(shape)
-            .shadow(
-                elevation = elevation,
-                shape = shape,
-                ambientColor = Color.Black.copy(alpha = shadowAlpha),
-                spotColor = Color.Black.copy(alpha = shadowAlpha)
-            )
             .border(1.dp, Color.White.copy(alpha = borderAlpha), shape),
         shape = shape,
         elevation = CardDefaults.cardElevation(defaultElevation = 0.dp),
